requirements.txt
fastapi==0.109.0
uvicorn==0.26.0
psycopg==3.1.9
psycopg-pool==3.1.7
pydantic-settings==2.1.0
psycopg-binary==3.1.17
typer==0.9.0
pydantic==2.5.3
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart==0.0.5
